refactoring and add PaymentButton test
