+++ ts/common/versionInfo/store/reducers/__mock__/ioVersionInfo.ts
@@ -1,0 +1,16 @@
+import { IOVersionInfo } from "../../../types/IOVersionInfo";
+
+export const mockIoVersionInfo: IOVersionInfo = {
+  min_app_version: {
+    ios: "0.0.0",
+    android: "0.0.0"
+  },
+  latest_released_app_version: {
+    ios: "0.0.0",
+    android: "0.0.0"
+  },
+  rollout_app_version: {
+    ios: "0.0.0",
+    android: "0.0.0"
+  }
+};
